feat: add new struct of tabs
